core/requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
asteval==1.0.8
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
boto3==1.42.59
botocore==1.42.59
bw-migrations==0.2
bw2calc==2.4.0
bw2data==4.5.4
bw2io==0.9.14
bw2parameters==1.1.0
bw_processing==1.0
cached-property==2.0.1
certifi==2026.2.25
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
constructive_geometries==1.0
country_converter==1.3.2
cryptography==46.0.5
dataflows-tabulator==1.54.3
datapackage==1.15.4
deepdiff==7.0.1
Deprecated==1.3.1
Django==5.2.11
django-allauth==65.14.3
django-countries==8.2.0
django-decouple==2.1
django-unfold==0.81.0
docopt==0.6.2
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
fsspec==2026.2.0
greenlet==3.3.2
idna==3.11
ijson==3.5.0
isodate==0.7.2
jmespath==1.1.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
linear-tsv==1.1.0
llvmlite==0.46.0
loguru==0.7.3
lxml==6.0.2
matrix_utils==0.7.0
morefs==0.2.2
mrio-common-metadata==0.2.1
numba==0.64.0
numpy==1.26.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==26.0
pandas==2.3.3
peewee==3.16.3
pillow==12.1.1
Pint==0.24.4
platformdirs==4.9.2
prettytable==3.17.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyecospold==4.0.1
PyJWT==2.11.0
pyparsing==3.3.2
PyPrind==2.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
pytz==2025.2
pyxlsb==1.0.10
PyYAML==6.0.3
randonneur==0.7.2
randonneur_data==0.7.2
RapidFuzz==3.14.3
rdflib==7.6.0
referencing==0.37.0
requests==2.32.5
rfc3986==2.0.0
rpds-py==0.30.0
s3transfer==0.16.0
scipy==1.15.3
six==1.17.0
snowflake-id==1.0.2
SPARQLWrapper==2.0.0
sparse==0.17.0
SQLAlchemy==2.0.47
sqlparse==0.5.5
stats_arrays==1.0
structlog==25.5.0
tableschema==1.21.0
toolz==1.1.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unicodecsv==0.14.1
urllib3==2.6.3
voluptuous==0.16.0
wcwidth==0.6.0
whitenoise==6.12.0
win32_setctime==1.2.0
wrapt==2.1.1
wurst==0.5.3
xarray==2025.6.1
xlrd==2.0.2
xlsxwriter==3.2.9
